refactor: 커스텀 예외 구조 변경
