+++ src/main/java/woowacourse/shoppingcart/dao/CustomerDao.java
@@ -9,7 +9,7 @@ import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.stereotype.Repository;
 import woowacourse.shoppingcart.domain.Customer;
-import woowacourse.shoppingcart.exception.notfound.InvalidCustomerException;
+import woowacourse.shoppingcart.exception.badrequest.InvalidCustomerException;
 import woowacourse.shoppingcart.exception.notfound.NotFoundCustomerException;
 
 @Repository
