+++ src/main/java/woowacourse/shoppingcart/dao/ProductDao.java
@@ -8,7 +8,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.stereotype.Repository;
 import woowacourse.shoppingcart.domain.Product;
-import woowacourse.shoppingcart.exception.notfound.InvalidProductException;
+import woowacourse.shoppingcart.exception.badrequest.InvalidProductException;
 
 @Repository
 public class ProductDao {
