+++ src/main/java/woowacourse/shoppingcart/exception/badrequest/InvalidProductException.java
@@ -1,6 +1,4 @@
-package woowacourse.shoppingcart.exception.notfound;
-
-import woowacourse.shoppingcart.exception.badrequest.BadRequestException;
+package woowacourse.shoppingcart.exception.badrequest;
 
 public class InvalidProductException extends BadRequestException {
